week2-job-scorer/requirements.txt
# Azure AI SDKs
azure-ai-documentintelligence>=1.0.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.8.0
openai>=1.42.0

# Prompt templating
jinja2>=3.1.0

# Structured output schema
pydantic>=2.8.0

# Evaluation
scikit-learn>=1.4.0

# CLI output
tabulate>=0.9.0

# UI
streamlit>=1.32.0

# Web scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0

# HTTP client for job search API
requests>=2.31.0

# Testing
pytest>=8.0.0
